refactor(ai): replaced custom xxhash64 with Bun.hash.xxHash64
- Removed the pure TypeScript xxhash64 implementation and its tests.
- Switched cch attestation to use the native Bun.hash.xxHash64 built-in.
- Migrated reference-value test cases into the alignment test suite.
